requirements.txt
--find-links https://download.pytorch.org/whl/cu118/torch_stable.html
moviepy
pydub
torch==2.0.0+cu118
torchvision==0.15.1+cu118
torchaudio==2.0.1+cu118
openai-whisper